Refactor _layout.tsx to _layout.jsx for Clerk integration
