+++ tsconfig.json
@@ -13,5 +13,5 @@
     "**/*.tsx",
     ".expo/types/**/*.ts",
     "expo-env.d.ts"
-, "app/_layout.tsx"  ]
+, "app/_layout.jsx"  ]
 }
